What are titles of books written by Krugman?
corrected ontology bugs, simplifying and correcting RDFTypes
reduce repetition in looking through sources
moved text to question answering
enhanced table processing for questions

+++ identifySpeakerGroups.cpp
@@ -2645,7 +2645,7 @@ void Source::eraseAliasesAndReplacementsInSpeakerGroups(void)
 			(sg+1)->sgBegin=sg->sgBegin;
 			(sg+1)->conversationalQuotes+=sg->conversationalQuotes;
 			(sg+1)->dnSpeakers.insert(sg->dnSpeakers.begin(),sg->dnSpeakers.end());
-			int numSG = sg - speakerGroups.begin();
+			int numSG = (int)(sg - speakerGroups.begin());
 			for (auto &tsg:speakerGroups)
 				if (tsg.previousSubsetSpeakerGroup >= numSG)
 					tsg.previousSubsetSpeakerGroup--;
